Another portion of fixes of tests.

+++ client/src/main/java/com/orientechnologies/orient/client/remote/message/OBeginTransaction38Request.java
@@ -25,14 +25,14 @@ import java.util.Map;
 
 public class OBeginTransaction38Request implements OBinaryRequest<OBeginTransactionResponse> {
 
-  private int txId;
+  private long txId;
   private boolean usingLog;
   private boolean hasContent;
   private List<ORecordOperationRequest> operations;
   private List<IndexChange> indexChanges;
 
   public OBeginTransaction38Request(
-      ODatabaseSessionInternal session, int txId,
+      ODatabaseSessionInternal session, long txId,
       boolean hasContent,
       boolean usingLog,
       Iterable<ORecordOperation> operations,
@@ -87,7 +87,7 @@ public class OBeginTransaction38Request implements OBinaryRequest<OBeginTransact
     // from 3.0 the the serializer is bound to the protocol
     ORecordSerializerNetworkV37Client serializer = ORecordSerializerNetworkV37Client.INSTANCE;
 
-    network.writeInt(txId);
+    network.writeLong(txId);
     network.writeBoolean(hasContent);
     network.writeBoolean(usingLog);
     if (hasContent) {
@@ -157,7 +157,7 @@ public class OBeginTransaction38Request implements OBinaryRequest<OBeginTransact
     return indexChanges;
   }
 
-  public int getTxId() {
+  public long getTxId() {
     return txId;
   }
