+++ client/src/main/java/com/orientechnologies/orient/client/remote/message/OCommit38Request.java
@@ -28,7 +28,7 @@ import java.util.Map;
  */
 public class OCommit38Request implements OBinaryRequest<OCommit37Response> {
 
-  private int txId;
+  private long txId;
   private boolean hasContent;
   private boolean usingLog;
   private List<ORecordOperationRequest> operations;
@@ -38,7 +38,7 @@ public class OCommit38Request implements OBinaryRequest<OCommit37Response> {
   }
 
   public OCommit38Request(
-      ODatabaseSessionInternal session, int txId,
+      ODatabaseSessionInternal session, long txId,
       boolean hasContent,
       boolean usingLong,
       Iterable<ORecordOperation> operations,
@@ -90,7 +90,7 @@ public class OCommit38Request implements OBinaryRequest<OCommit37Response> {
       OStorageRemoteSession session) throws IOException {
     // from 3.0 the the serializer is bound to the protocol
     ORecordSerializerNetworkV37Client serializer = ORecordSerializerNetworkV37Client.INSTANCE;
-    network.writeInt(txId);
+    network.writeLong(txId);
     network.writeBoolean(hasContent);
     network.writeBoolean(usingLog);
     if (hasContent) {
@@ -151,7 +151,7 @@ public class OCommit38Request implements OBinaryRequest<OCommit37Response> {
     return "Commit";
   }
 
-  public int getTxId() {
+  public long getTxId() {
     return txId;
   }
